backend/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
Fiona==1.9.3
Flask==2.2.3
Flask-Cors==3.0.10
folium==0.13.0
fonttools==4.39.3
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
gpxpy==1.5.0
graphviz==0.20.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
packaging==23.1
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.15.1
pyparsing==3.0.9
pyproj==3.3.1
pyshp==2.3.1
python-dateutil==2.8.2
PyTrack-lib==2.0.7
pytz==2023.3
pyzmq==25.0.2
requests==2.28.2
scikit-learn==1.1.3
scipy==1.10.1
Shapely==1.8.5.post1
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3
traitlets==5.9.0
urllib3==1.26.15
waitress==2.1.2
wcwidth==0.2.6
Werkzeug==2.2.3
